requirements.txt
# Essencial
pillow>=10.0.0
numpy>=1.24.0
PyMuPDF>=1.23.0
nltk>=3.8.1
